Add relation between users and orders

+++ src/orders/orders.service.ts
@@ -1,5 +1,6 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { User } from 'src/auth/user.entity';
 import { CreateOrderDto } from './dto/create-order.dto';
 import { GetOrdersFilterDto } from './dto/get-orders-filter.dto';
 import { OrderStatus } from './order-status.enum';
@@ -13,31 +14,40 @@ export class OrdersService {
     private orderRepository: OrderRepository,
   ) {}
 
-  getOrders(filterDto: GetOrdersFilterDto): Promise<Order[]> {
-    return this.orderRepository.getOrders(filterDto);
+  getOrders(filterDto: GetOrdersFilterDto, user: User): Promise<Order[]> {
+    return this.orderRepository.getOrders(filterDto, user);
   }
 
-  async getOrderById(id: number): Promise<Order> {
-    const found = await this.orderRepository.findOne(id);
+  async getOrderById(id: number, user: User): Promise<Order> {
+    const found = await this.orderRepository.findOne({
+      where: { id, userId: user.id },
+    });
     if (!found) {
       throw new NotFoundException(`Order ID "${id}" not found!`);
     }
     return found;
   }
 
-  async createOrder(createOrderDto: CreateOrderDto): Promise<Order> {
-    return this.orderRepository.createOrder(createOrderDto);
+  async createOrder(
+    createOrderDto: CreateOrderDto,
+    user: User,
+  ): Promise<Order> {
+    return this.orderRepository.createOrder(createOrderDto, user);
   }
 
-  async updateOrderStatus(id: number, status: OrderStatus): Promise<Order> {
-    const order = await this.getOrderById(id);
+  async updateOrderStatus(
+    id: number,
+    status: OrderStatus,
+    user: User,
+  ): Promise<Order> {
+    const order = await this.getOrderById(id, user);
     order.status = status;
     order.save();
     return order;
   }
 
-  async deleteOrder(id: number): Promise<void> {
-    const result = await this.orderRepository.delete(id);
+  async deleteOrder(id: number, user: User): Promise<void> {
+    const result = await this.orderRepository.delete({ id, userId: user.id });
     if (result.affected === 0) {
       throw new NotFoundException(`Order ID "${id}" not found!`);
     }
